requirements.txt
numpy>=1.24
PyYAML>=6.0
matplotlib>=3.7
seaborn>=0.12
Pillow>=10.0
torch>=2.2
torchvision>=0.17
